[ERD] Modified all price types in database design to DOUBLE; [db.database] Modified all prices in DB builder/checker to DOUBLE; [reservation.dto] Modified "Payment.amount" and "RoomPrice.price" to type 'Double'; [reservation.delegate] Modified "IReserve" and interface implementation of methods "cancelReservation(..)" and "cancelReservedRoom(..)" return types to 'Double'; [TEST] Modified prices in "TestDataDBQueries" to Double types.

+++ services/src/main/java/bbk_beam/mtRooms/db/database/DatabaseChecker.java
@@ -352,7 +352,7 @@ class DatabaseChecker {
                 }
                 if (row.get("name").equals("price")) {
                     checked++;
-                    ColProperty expected = new ColProperty("RoomPrice", "price", "INTEGER", true, null, 0);
+                    ColProperty expected = new ColProperty("RoomPrice", "price", "DOUBLE", true, null, 0);
                     if (!checkColumn(expected, row))
                         ok_flag = false;
                 }
@@ -652,7 +652,7 @@ class DatabaseChecker {
                 }
                 if (row.get("name").equals("amount")) {
                     checked++;
-                    ColProperty expected = new ColProperty("Payment", "amount", "INTEGER", true, null, 0);
+                    ColProperty expected = new ColProperty("Payment", "amount", "DOUBLE", true, null, 0);
                     if (!checkColumn(expected, row))
                         ok_flag = false;
                 }
